move regression case to mtmv_p0/ivm
