+++ regression-test/data/mtmv_p0/ivm/test_ivm_agg_1.out
@@ -1,0 +1,56 @@
+-- This file is automatically generated. You should know what you did if you want to edit this
+-- !agg_after_first_complete --
+1	1	10
+2	1	20
+3	1	30
+
+-- !agg_after_complete --
+1	1	15
+2	1	25
+3	1	30
+4	1	40
+
+-- !scalar_after_first_complete --
+3	60	20	3
+
+-- !scalar_after_final_complete --
+4	105	26.25	4
+
+-- !minmax_after_first_complete --
+1	10	10
+2	20	20
+3	30	30
+
+-- !minmax_after_incremental --
+1	5	10
+2	20	20
+3	30	30
+4	40	40
+
+-- !minmax_after_final_complete --
+1	5	5
+2	20	20
+3	30	30
+4	40	40
+
+-- !minmax_op_after_complete --
+10	30	3
+
+-- !minmax_op_after_recovery --
+10	35	5
+
+-- !null_after_complete --
+1	30	15	2	10	20	2
+2	30	30	1	30	30	2
+3	\N	\N	0	\N	\N	1
+
+-- !null_after_final_complete --
+1	30	15	2	10	20	3
+2	30	30	1	30	30	2
+3	50	50	1	50	50	2
+
+-- !null_after_delete_complete --
+1	30	15	2	10	20	3
+2	70	35	2	30	40	3
+3	50	50	1	50	50	2
+
