+++ regression-test/data/mtmv_p0/ivm/test_ivm_agg_2.out
@@ -1,0 +1,46 @@
+-- This file is automatically generated. You should know what you did if you want to edit this
+-- !allnull_after_complete --
+\N	\N	0	\N	\N	2
+
+-- !allnull_after_insert_complete --
+100	100	1	100	100	3
+
+-- !allnull_after_null_insert_complete --
+100	100	1	100	100	4
+
+-- !grpdel_after_complete --
+1	2	30
+2	1	30
+
+-- !grpdel_after_incremental --
+1	5	100
+
+-- !grpdel_after_final_complete --
+1	3	70
+2	1	30
+
+-- !grpdel_after_resurrect_complete --
+1	3	70
+2	2	80
+
+-- !scalardel_after_complete --
+2	30	10	20
+
+-- !scalardel_countsum_after_complete --
+2	30
+
+-- !scalardel_countsum_after_delete_complete --
+3	129
+
+-- !scalardel_countsum_after_allempty_complete --
+4	129
+
+-- !scalardel_countsum_after_recovery_complete --
+6	239
+
+-- !maxdel_after_complete --
+10	30	3
+
+-- !maxdel_after_recovery --
+10	30	5
+
