+++ regression-test/data/mtmv_p0/ivm/test_ivm_agg_3.out
@@ -1,0 +1,49 @@
+-- This file is automatically generated. You should know what you did if you want to edit this
+-- !multikey_after_complete --
+1	a	2	30
+1	b	1	30
+2	a	1	40
+
+-- !multikey_after_insert_complete --
+1	a	3	80
+1	b	1	30
+2	a	1	40
+2	b	1	60
+
+-- !multikey_after_delete_incremental --
+1	a	6	160
+2	a	3	150
+2	b	2	120
+
+-- !multikey_after_final_complete --
+1	a	3	80
+1	b	1	30
+2	a	2	110
+2	b	1	60
+
+-- !multiagg_after_complete --
+1	30	300	10	200	2
+2	30	300	30	300	1
+
+-- !multiagg_after_insert_complete --
+1	45	450	10	200	3
+2	65	550	30	300	2
+
+-- !negval_after_complete --
+2	20
+
+-- !negval_after_negative_complete --
+3	-30
+
+-- !negval_after_zero_complete --
+4	0
+
+-- !emptydelta_unchanged --
+2	30
+
+-- !types_after_complete --
+25	8.333333333333334	250.50	83.5000	7.5	2.5	-5	100.75	3
+
+-- !types_after_insert_complete --
+25	6.25	325.50	81.3750	11.5	2.875	-5	100.75	4
+
