--- regression-test/data/mtmv_p0/test_ivm_agg_mtmv.out
@@ -1,149 +1,0 @@
--- This file is automatically generated. You should know what you did if you want to edit this
--- !agg_after_first_complete --
-1	1	10
-2	1	20
-3	1	30
-
--- !agg_after_complete --
-1	1	15
-2	1	25
-3	1	30
-4	1	40
-
--- !scalar_after_first_complete --
-3	60	20	3
-
--- !scalar_after_final_complete --
-4	105	26.25	4
-
--- !minmax_after_first_complete --
-1	10	10
-2	20	20
-3	30	30
-
--- !minmax_after_incremental --
-1	5	10
-2	20	20
-3	30	30
-4	40	40
-
--- !minmax_after_final_complete --
-1	5	5
-2	20	20
-3	30	30
-4	40	40
-
--- !minmax_op_after_complete --
-10	30	3
-
--- !minmax_op_after_recovery --
-10	35	5
-
--- !null_after_complete --
-1	30	15	2	10	20	2
-2	30	30	1	30	30	2
-3	\N	\N	0	\N	\N	1
-
--- !null_after_final_complete --
-1	30	15	2	10	20	3
-2	30	30	1	30	30	2
-3	50	50	1	50	50	2
-
--- !null_after_delete_complete --
-1	30	15	2	10	20	3
-2	70	35	2	30	40	3
-3	50	50	1	50	50	2
-
--- !allnull_after_complete --
-\N	\N	0	\N	\N	2
-
--- !allnull_after_insert_complete --
-100	100	1	100	100	3
-
--- !allnull_after_null_insert_complete --
-100	100	1	100	100	4
-
--- !grpdel_after_complete --
-1	2	30
-2	1	30
-
--- !grpdel_after_incremental --
-1	5	100
-
--- !grpdel_after_final_complete --
-1	3	70
-2	1	30
-
--- !grpdel_after_resurrect_complete --
-1	3	70
-2	2	80
-
--- !scalardel_after_complete --
-2	30	10	20
-
--- !scalardel_countsum_after_complete --
-2	30
-
--- !scalardel_countsum_after_delete_complete --
-3	129
-
--- !scalardel_countsum_after_allempty_complete --
-4	129
-
--- !scalardel_countsum_after_recovery_complete --
-6	239
-
--- !maxdel_after_complete --
-10	30	3
-
--- !maxdel_after_recovery --
-10	30	5
-
--- !multikey_after_complete --
-1	a	2	30
-1	b	1	30
-2	a	1	40
-
--- !multikey_after_insert_complete --
-1	a	3	80
-1	b	1	30
-2	a	1	40
-2	b	1	60
-
--- !multikey_after_delete_incremental --
-1	a	6	160
-2	a	3	150
-2	b	2	120
-
--- !multikey_after_final_complete --
-1	a	3	80
-1	b	1	30
-2	a	2	110
-2	b	1	60
-
--- !multiagg_after_complete --
-1	30	300	10	200	2
-2	30	300	30	300	1
-
--- !multiagg_after_insert_complete --
-1	45	450	10	200	3
-2	65	550	30	300	2
-
--- !negval_after_complete --
-2	20
-
--- !negval_after_negative_complete --
-3	-30
-
--- !negval_after_zero_complete --
-4	0
-
--- !emptydelta_unchanged --
-2	30
-
--- !types_after_complete --
-25	8.333333333333334	250.50	83.5000	7.5	2.5	-5	100.75	3
-
--- !types_after_insert_complete --
-25	6.25	325.50	81.3750	11.5	2.875	-5	100.75	4
-
